refactor(consensus): reduce the nameservice through the generic seam
Deletes the second openraft adapter. The nameservice's state machine
was ~450 lines of last-applied bookkeeping, membership storage, and
snapshot persistence that `fluree-raft-core` already owns generically,
wrapped around the part that is actually about naming ledgers.

That part now lives in `raft::app`, split the way the seam splits it:
`NameServiceApp` is the pure reduction, `NameServiceObserver` is the
effects — event bus, waiter resolution, staged receipts, content-store
releases, ledger-cache watermark. `raft::state_machine_adapter` is their
composition, kept at its historical path so imports resolve.

Two orderings were structural in the old `apply` and are now explicit,
because both are easy to lose and neither had a test.

`publish` runs in two phases: every commit-head watermark reaches the
ledger cache before any event reaches the bus. The watermark is a
synchronous lossless write and the bus is bounded and lossy, so a cache
lookup racing an apply must already compare against the new head. Doing
it per-event would order each ledger against its own event but let one
ledger's event overtake another's watermark.

Within phase two, effects run in the order they were recorded, so a
waiter bind lands before the terminal command that resolves it — a
single apply batch can carry both, and a pass that handled terminals
first would strand the proposer on work that had already succeeded.

The snapshot format is unchanged: `encode_snapshot` returns bare
postcard and the generic adapter stores it verbatim, so deployed
clusters read and write the same bytes. `codec`'s versioned envelope is
offered, not imposed, which is what makes that possible.

Test scaffolding that wrote state directly now goes through the real
paths — a demotion through `SetWorkerEligibility`, a seeded head through
storage — since `apply` is the only writer. `ReadOnlyState::view_of`
lets a test own state and hand out the read-only view consumers see; it
only ever narrows.

Verified both orderings fail when broken: binding after terminals
strands the waiter, and dropping event-carried heads from phase one
leaves the cache stale.
