+++ docs/design/raft-command-queue.md
@@ -20,7 +20,9 @@ The Raft consensus crate (`fluree-db-consensus/src/raft/`) is structured as a se
 | -------------------------- | -------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------- |
 | `Command`, `Response`      | `state_machine.rs`                                       | The log entry types. ~20 variants spanning transaction flow, ledger lifecycle, and metadata.                |
 | `NameServiceState`         | `state_machine.rs`                                       | The replicated in-memory state: branch heads, ledger registry, per-branch queues, idempotency cache.        |
-| `StateMachineAdapter`      | `state_machine_adapter.rs`                               | openraft's `RaftStateMachine` impl. Applies entries, takes/installs snapshots, resolves waiters.            |
+| `NameServiceApp`           | `app.rs`                                                 | The pure reduction: routes a `Command` through `state_machine::apply`, mirrors membership, owns the snapshot codec. |
+| `NameServiceObserver`      | `app.rs`                                                 | The effects: event bus, waiter resolution, staged receipts, releases, ledger-cache watermark. Captured under the state lock, published after it drops. |
+| `StateMachineAdapter`      | `state_machine_adapter.rs`                               | The two above composed with `fluree_raft_core::state_machine::StateMachineAdapter`, which owns last-applied, membership, and snapshot persistence. |
 | `LogStore`, `SnapshotStore`| `fluree-raft-core` `log_adapter.rs`, `storage/{fs,memory}.rs` | openraft's `RaftLogStorage`/`RaftLogReader`/`RaftSnapshotBuilder`. Local-disk persistence. The backends are openraft-free — they store opaque `Vec<u8>` payloads. |
 | `HttpRaftNetworkFactory`   | `fluree-raft-core` `network.rs`                          | Inter-node RPC (`/raft/vote`, `/raft/append-entries`, `/raft/install-snapshot`) over HTTP. Generic over the type config; `network.rs` here keeps only the nameservice-specific `NetworkConfig` fields. |
 | `RaftAdmin` / `/cluster/*` | `fluree-raft-core` `admin.rs`                            | Operator-facing membership endpoints (`initialize`, `add-learner`, `change-membership`, `status`).         |
