+++ fluree-db-consensus/src/raft/state_machine_adapter.rs
@@ -1,1700 +1,25 @@
-//! openraft state-machine adapter.
+//! The nameservice's openraft state machine.
 //!
-//! Holds the [`NameServiceState`] plus the bookkeeping openraft
-//! requires (last-applied log id, current membership) and routes
-//! `apply` through our pure [`state_machine::apply`]. Snapshot
-//! persistence happens via the [`RaftStorage::snapshots`] handle.
+//! Both halves now live elsewhere: the generic bookkeeping — last
+//! applied, membership, snapshot persistence, the persist-before-swap
+//! ordering — in [`fluree_raft_core::state_machine::StateMachineAdapter`],
+//! and the nameservice-specific reduction and effects in
+//! [`super::app`]. This module is the composition of the two, kept at
+//! its historical path so existing imports resolve.
 //!
-//! The post-apply release processing (driving content-store deletes
-//! from `Response::Applied.release`) is a separate concern that lives
-//! at the wrapper above this adapter — see phase 4 of the design.
-//! For now apply just returns the `Response` as-is.
-
-use crate::raft::log_adapter::{from_openraft_log_id, to_openraft_log_id};
-use crate::raft::staged_receipt::{AppliedReceipt, StagedReceiptMap};
-use crate::raft::state_machine::{self, Command, NameServiceState, RefKey, Response};
-use crate::raft::storage::{
-    RaftSnapshotStore, RaftStorage, SnapshotId as OurSnapshotId, SnapshotMeta as OurSnapshotMeta,
-};
-use crate::raft::waiter::{AbortReason, WaiterMap};
-use crate::raft::{ClusterNode, NodeId, TypeConfig};
-use fluree_db_api::LedgerManager;
-use fluree_db_core::ledger_id::format_ledger_id;
-use fluree_db_core::ContentId;
-use fluree_db_nameservice::{LedgerEventBus, NameServiceEvent};
-use openraft::storage::{RaftSnapshotBuilder, RaftStateMachine};
-use openraft::{
-    AnyError, Entry, EntryPayload, ErrorSubject, ErrorVerb, LogId, Snapshot, SnapshotMeta,
-    StorageError, StorageIOError, StoredMembership,
-};
-use std::collections::BTreeSet;
-use std::io::Cursor;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::{Arc, OnceLock};
-use tokio::sync::mpsc::UnboundedSender;
-use tokio::sync::RwLock;
-
-/// Handle to the replicated nameservice state used by the apply
-/// path. Cloning is cheap (`Arc` clone). Sharing this with a
-/// [`RaftNameService`](crate::raft::nameservice::RaftNameService)
-/// lets read-only consumers see committed state without going
-/// through openraft's RPC surface.
-pub type SharedState = Arc<RwLock<NameServiceState>>;
-
-fn io_err<S: ToString>(
-    verb: ErrorVerb,
-    subject: ErrorSubject<NodeId>,
-    source: S,
-) -> StorageError<NodeId> {
-    StorageError::IO {
-        source: StorageIOError::new(subject, verb, AnyError::error(source.to_string())),
-    }
-}
-
-fn read_state_err<S: ToString>(source: S) -> StorageError<NodeId> {
-    io_err(ErrorVerb::Read, ErrorSubject::StateMachine, source)
-}
-
-fn write_state_err<S: ToString>(source: S) -> StorageError<NodeId> {
-    io_err(ErrorVerb::Write, ErrorSubject::StateMachine, source)
-}
-
-fn snapshot_err<S: ToString>(verb: ErrorVerb, source: S) -> StorageError<NodeId> {
-    StorageError::IO {
-        source: StorageIOError::new(
-            ErrorSubject::Snapshot(None),
-            verb,
-            AnyError::error(source.to_string()),
-        ),
-    }
-}
-
-/// openraft state-machine adapter wrapping an `Arc<S: RaftStorage>`.
-///
-/// Holds the in-memory [`NameServiceState`] (shared via [`SharedState`]
-/// so a [`RaftNameService`](crate::raft::nameservice::RaftNameService)
-/// can read the same committed state) plus the bookkeeping openraft
-/// needs (last-applied log id, current membership). Snapshot
-/// reads/writes go through `S::SnapshotStore`.
-pub struct StateMachineAdapter<S>
-where
-    S: RaftStorage,
-{
-    state: SharedState,
-    last_applied: Option<LogId<NodeId>>,
-    last_membership: StoredMembership<NodeId, ClusterNode>,
-    storage: Arc<S>,
-    /// Monotonic counter for snapshot ids — combined with the
-    /// last-applied index for uniqueness across rebuilds.
-    snapshot_counter: AtomicU64,
-    /// When set, the adapter emits [`NameServiceEvent`]s on this bus
-    /// after each successful apply.
-    event_bus: Option<Arc<LedgerEventBus>>,
-    /// When set, the adapter resolves queue waiters parked on this
-    /// node after each successful queue-related apply. Per-process
-    /// scope — leader transitions strand waiters from the prior
-    /// leader, which the proposer recovers from with timeout +
-    /// idempotency-keyed retry.
-    waiter_map: Option<Arc<WaiterMap>>,
-    /// Companion to [`waiter_map`](Self::waiter_map): the worker
-    /// stashes per-op staging detail here before proposing
-    /// `ApplyHead`; the adapter takes it during waiter resolution
-    /// and sends it through as part of [`AppliedReceipt`]. Absent
-    /// entry → fall back to [`AppliedReceipt::Minimal`] (the
-    /// stranded-on-former-leader path).
-    staged_receipts: Option<Arc<StagedReceiptMap>>,
-    /// When set, the adapter forwards `(ledger_id, request_cid)` pairs
-    /// surfaced by `Response::EvictionApplied`, `Purged`,
-    /// `BranchDropped`, and `HeadReset` to a per-node release task.
-    /// The task owns the `Fluree` handle and calls
-    /// `content_store(&ledger_id).release(&cid)` — running on every
-    /// node so admin clears and idempotency eviction don't orphan
-    /// envelope blobs on followers.
-    release_tx: Option<UnboundedSender<(String, ContentId)>>,
-    /// Late-bound handle to the api-layer ledger cache. The adapter
-    /// reports commit-head advances into its watermark
-    /// ([`LedgerManager::note_head_advance`]) inline during `apply`
-    /// and `install_snapshot`, before broadcasting events — a
-    /// lossless channel, unlike the bounded event bus, and a bare
-    /// memory write, so apply latency is unaffected. A cell rather
-    /// than a builder field because the cache is constructed after
-    /// the adapter: bootstrap clones the cell out, and server
-    /// assembly fills it once `Fluree` exists. Empty cell → applies
-    /// proceed without reporting.
-    ledger_manager: Arc<OnceLock<Arc<LedgerManager>>>,
-}
-
-impl<S> StateMachineAdapter<S>
-where
-    S: RaftStorage,
-{
-    /// Construct an adapter with a freshly-allocated [`SharedState`]
-    /// and no restored persistence. Tests and fresh-cluster bootstrap
-    /// only — production restart paths must use [`Self::open`] so
-    /// committed state from the last snapshot is loaded before
-    /// openraft replays the post-snapshot log tail.
-    pub fn new(storage: Arc<S>) -> Self {
-        Self::with_state(storage, Arc::new(RwLock::new(NameServiceState::default())))
-    }
-
-    /// Construct an adapter sharing the provided state handle. Use
-    /// this when the same state must be visible to a `RaftNameService`
-    /// constructed alongside the adapter.
-    pub fn with_state(storage: Arc<S>, state: SharedState) -> Self {
-        Self {
-            state,
-            last_applied: None,
-            last_membership: StoredMembership::default(),
-            storage,
-            snapshot_counter: AtomicU64::new(0),
-            event_bus: None,
-            waiter_map: None,
-            staged_receipts: None,
-            release_tx: None,
-            ledger_manager: Arc::new(OnceLock::new()),
-        }
-    }
-
-    /// Clone out the late-binding ledger-cache cell so the
-    /// integration layer can fill it after `Fluree` exists.
-    pub fn ledger_manager_cell(&self) -> Arc<OnceLock<Arc<LedgerManager>>> {
-        Arc::clone(&self.ledger_manager)
-    }
-
-    /// Construct an adapter and restore state + `last_applied` +
-    /// `last_membership` from `storage`'s current snapshot if one
-    /// exists. This is the production restart path: without it, the
-    /// adapter would boot with `last_applied = None` and openraft
-    /// would replay the log from index 0 — but after a snapshot +
-    /// log purge (the default `SnapshotPolicy`) the pre-snapshot log
-    /// entries are gone, so committed nameservice state would be
-    /// silently lost. With the snapshot restored here, openraft's
-    /// replay starts at the snapshot's `last_applied + 1`.
-    ///
-    /// Note this restores from the *snapshot's* last-applied, not
-    /// the last entry applied before the crash: any entry committed
-    /// after the snapshot but before the crash is re-applied on
-    /// recovery, so its post-apply side effects (CAS releases,
-    /// event-bus emissions) do re-fire. That's benign — `release`
-    /// is idempotent and event subscribers tolerate duplicates —
-    /// but it is not "exactly once."
-    pub async fn open(storage: Arc<S>) -> Result<Self, StorageError<NodeId>> {
-        let mut adapter = Self::new(storage);
-        adapter.restore_from_snapshot().await?;
-        Ok(adapter)
-    }
-
-    /// Load the latest snapshot from storage (if any) into this
-    /// adapter's in-memory state. No-op when storage has no snapshot
-    /// (fresh node). Idempotent — repeated calls overwrite with the
-    /// same snapshot.
-    async fn restore_from_snapshot(&mut self) -> Result<(), StorageError<NodeId>> {
-        let current = self
-            .storage
-            .snapshots()
-            .current()
-            .await
-            .map_err(|e| snapshot_err(ErrorVerb::Read, e))?;
-        let Some((our_meta, bytes)) = current else {
-            return Ok(());
-        };
-        let restored_state = NameServiceState::from_snapshot(&bytes).map_err(read_state_err)?;
-        let last_membership: StoredMembership<NodeId, ClusterNode> =
-            postcard::from_bytes(&our_meta.membership).map_err(read_state_err)?;
-        *self.state.write().await = restored_state;
-        self.last_applied = our_meta.last_applied.map(to_openraft_log_id);
-        self.last_membership = last_membership;
-        Ok(())
-    }
-
-    /// Set the [`LedgerEventBus`] this adapter emits commit/index
-    /// events on.
-    pub fn with_event_bus(mut self, event_bus: Arc<LedgerEventBus>) -> Self {
-        self.event_bus = Some(event_bus);
-        self
-    }
-
-    /// Set the [`WaiterMap`] this adapter resolves after each
-    /// queue-related apply. Pair it with the same handle the
-    /// [`QueuedTransactor`](super::queued_transactor::QueuedTransactor)
-    /// registers waiters on.
-    pub fn with_waiter_map(mut self, waiter_map: Arc<WaiterMap>) -> Self {
-        self.waiter_map = Some(waiter_map);
-        self
-    }
-
-    /// Set the [`StagedReceiptMap`] the adapter reads from when
-    /// constructing an [`AppliedReceipt`] for a resolved waiter.
-    /// Pair it with the same handle the per-branch
-    /// [`Worker`](super::commit_worker::Worker) stashes per-op staging
-    /// detail into.
-    pub fn with_staged_receipts(mut self, staged_receipts: Arc<StagedReceiptMap>) -> Self {
-        self.staged_receipts = Some(staged_receipts);
-        self
-    }
-
-    /// Set the channel the adapter forwards CAS release work to.
-    /// Pair the receiver with a per-node task that owns the `Fluree`
-    /// handle and calls `content_store(&ledger_id).release(&cid)`.
-    pub fn with_release_sender(mut self, tx: UnboundedSender<(String, ContentId)>) -> Self {
-        self.release_tx = Some(tx);
-        self
-    }
-
-    /// Borrow the shared state handle. Cheap clone (`Arc`).
-    pub fn shared_state(&self) -> SharedState {
-        Arc::clone(&self.state)
-    }
-}
-
-/// What the waiter-map should do for a single `(Command, Response)`
-/// pair. Computed under the apply lock, executed after it drops so
-/// subscribers (the parked waiters' senders' receivers) can't
-/// reenter apply.
-enum WaiterResolution {
-    /// This node applied an `EnqueueCommand`. If a local proposer
-    /// armed interest in its `request_cid`, bind that interest to the
-    /// `queue_id` the state machine just assigned.
-    ///
-    /// Ordered with the resolutions below rather than done inline,
-    /// because a single apply batch can carry both the enqueue and its
-    /// terminal command — the bind has to land first.
-    ///
-    /// A follower has nothing armed, so this is a no-op there. That is
-    /// the whole reason a follower's waiter map stays empty.
-    Bind {
-        request_cid: ContentId,
-        queue_id: u64,
-    },
-    /// `ApplyHead` advanced the head — wake the parked transactor
-    /// with the new head identity.
-    Applied {
-        queue_id: u64,
-        commit_id: ContentId,
-        commit_t: i64,
-    },
-    /// `PoisonQueueEntry` recorded a terminal failure — wake the
-    /// transactor with the abort reason.
-    Aborted { queue_id: u64, reason: AbortReason },
-    /// A head-mutating admin command cleared the per-branch queue —
-    /// wake every parked transactor on that branch with the
-    /// matching abort reason.
-    AbortBranch {
-        ref_key: RefKey,
-        reason: AbortReason,
-    },
-}
-
-/// Translate an apply-path `(Command, Response)` pair into the
-/// matching waiter resolution, if any. Returns `None` for pairs that
-/// don't terminate a queue entry (every non-queue command, plus
-/// `QueueDesync` — the waiter for that queue_id has already been
-/// resolved by whichever earlier event popped it).
-fn waiter_resolution_for(cmd: &Command, response: &Response) -> Option<WaiterResolution> {
-    match (cmd, response) {
-        (
-            Command::EnqueueCommand(args),
-            Response::Enqueued { queue_id, .. } | Response::InFlight { queue_id, .. },
-        ) => Some(WaiterResolution::Bind {
-            request_cid: args.request_cid.clone(),
-            queue_id: *queue_id,
-        }),
-        (
-            Command::ApplyHead(args),
-            Response::HeadApplied {
-                commit_id,
-                commit_t,
-                ..
-            },
-        ) => Some(WaiterResolution::Applied {
-            queue_id: args.queue_id,
-            commit_id: commit_id.clone(),
-            commit_t: *commit_t,
-        }),
-        (
-            Command::PoisonQueueEntry(_),
-            Response::Poisoned {
-                queue_id, reason, ..
-            },
-        ) => Some(WaiterResolution::Aborted {
-            queue_id: *queue_id,
-            reason: AbortReason::Poisoned(reason.clone()),
-        }),
-        (
-            Command::DropBranch {
-                ledger_id, branch, ..
-            },
-            Response::BranchDropped { .. },
-        ) => Some(WaiterResolution::AbortBranch {
-            ref_key: RefKey::new(ledger_id, branch),
-            reason: AbortReason::BranchDropped,
-        }),
-        (
-            Command::PurgeBranch {
-                ledger_id, branch, ..
-            },
-            Response::Purged { .. },
-        ) => Some(WaiterResolution::AbortBranch {
-            ref_key: RefKey::new(ledger_id, branch),
-            reason: AbortReason::BranchPurged,
-        }),
-        (
-            Command::ResetHead {
-                ledger_id, branch, ..
-            },
-            Response::HeadReset { .. },
-        ) => Some(WaiterResolution::AbortBranch {
-            ref_key: RefKey::new(ledger_id, branch),
-            reason: AbortReason::BranchHeadReset,
-        }),
-        (
-            Command::RetractLedger {
-                ledger_id, branch, ..
-            },
-            Response::Retracted { .. },
-        ) => Some(WaiterResolution::AbortBranch {
-            ref_key: RefKey::new(ledger_id, branch),
-            reason: AbortReason::BranchRetracted,
-        }),
-        _ => None,
-    }
-}
-
-/// Translate an apply-path `(Command, Response)` pair into the
-/// matching [`NameServiceEvent`]. Returns `None` for pairs that
-/// don't advance head state — desyncs, no-ops, idempotency hits.
-fn event_for(cmd: &Command, response: &Response) -> Option<NameServiceEvent> {
-    match (cmd, response) {
-        (
-            Command::ApplyHead(args),
-            Response::HeadApplied {
-                commit_id,
-                commit_t,
-                ..
-            },
-        ) => Some(NameServiceEvent::LedgerCommitPublished {
-            ledger_id: format_ledger_id(&args.ledger_id, &args.branch),
-            commit_id: commit_id.clone(),
-            commit_t: *commit_t,
-        }),
-        (
-            Command::AdvanceIndexHead(args),
-            Response::IndexAdvanced {
-                index_t,
-                index_head,
-            },
-        ) => Some(NameServiceEvent::LedgerIndexPublished {
-            ledger_id: format_ledger_id(&args.ledger_id, &args.branch),
-            index_id: index_head.clone(),
-            index_t: *index_t,
-        }),
-        (Command::RetractLedger { .. }, Response::Retracted { ledger_id, .. })
-        | (Command::PurgeBranch { .. }, Response::Purged { ledger_id, .. })
-        | (Command::DropBranch { .. }, Response::BranchDropped { ledger_id, .. }) => {
-            Some(NameServiceEvent::LedgerRetracted {
-                ledger_id: ledger_id.clone(),
-            })
-        }
-        (Command::CreateBranch(_), Response::BranchCreated { ledger_id, head, t }) => {
-            Some(NameServiceEvent::LedgerCommitPublished {
-                ledger_id: ledger_id.clone(),
-                commit_id: head.clone(),
-                commit_t: *t,
-            })
-        }
-        _ => None,
-    }
-}
-
-/// Drain the `(ledger_id, request_cid)` pairs the state machine has
-/// flagged for content-store release. Covers idempotency eviction
-/// and the three head-mutating admin commands that clear pending
-/// queue entries. Returning `Vec` (not a `&[..]`) lets the adapter
-/// move the pairs into the release channel without an extra copy.
-fn drain_releases(response: &mut Response) -> Vec<(String, ContentId)> {
-    match response {
-        Response::EvictionApplied {
-            released_envelopes, ..
-        }
-        | Response::Purged {
-            released_envelopes, ..
-        }
-        | Response::BranchDropped {
-            released_envelopes, ..
-        }
-        | Response::HeadReset {
-            released_envelopes, ..
-        }
-        | Response::Retracted {
-            released_envelopes, ..
-        } => std::mem::take(released_envelopes),
-        // A keyless queue entry's envelope isn't held by any
-        // idempotency record, so `apply_head` / `apply_poison`
-        // release it as the entry retires.
-        Response::HeadApplied {
-            ledger_id,
-            released_envelope,
-            ..
-        }
-        | Response::Poisoned {
-            ledger_id,
-            released_envelope,
-            ..
-        } => released_envelope
-            .take()
-            .map(|cid| (ledger_id.clone(), cid))
-            .into_iter()
-            .collect(),
-        _ => Vec::new(),
-    }
-}
-
-impl<S> RaftStateMachine<TypeConfig> for StateMachineAdapter<S>
-where
-    S: RaftStorage,
-{
-    type SnapshotBuilder = SnapshotBuilder<S>;
-
-    async fn applied_state(
-        &mut self,
-    ) -> Result<(Option<LogId<NodeId>>, StoredMembership<NodeId, ClusterNode>), StorageError<NodeId>>
-    {
-        Ok((self.last_applied, self.last_membership.clone()))
-    }
-
-    async fn apply<I>(&mut self, entries: I) -> Result<Vec<Response>, StorageError<NodeId>>
-    where
-        I: IntoIterator<Item = Entry<TypeConfig>> + Send,
-        I::IntoIter: Send,
-    {
-        let mut responses = Vec::new();
-        let mut events = Vec::new();
-        let mut resolutions = Vec::new();
-        let mut releases: Vec<(String, ContentId)> = Vec::new();
-        {
-            let mut state = self.state.write().await;
-            for entry in entries {
-                let log_id = entry.log_id;
-                self.last_applied = Some(log_id);
-                match entry.payload {
-                    EntryPayload::Blank => responses.push(Response::NoOp),
-                    EntryPayload::Normal(cmd) => {
-                        let mut response =
-                            state_machine::apply(&mut state, cmd.clone(), log_id.index);
-                        if let Some(event) = event_for(&cmd, &response) {
-                            events.push(event);
-                        }
-                        if let Some(resolution) = waiter_resolution_for(&cmd, &response) {
-                            resolutions.push(resolution);
-                        }
-                        releases.extend(drain_releases(&mut response));
-                        responses.push(response);
-                    }
-                    EntryPayload::Membership(m) => {
-                        // Mirror the new voter set into the replicated
-                        // state machine so `apply_set_worker_eligibility`
-                        // can validate against it without threading
-                        // membership through the apply signature.
-                        //
-                        // Voters demoted before the change keep their
-                        // demotion if they remain in the new voter
-                        // set; newly-added voters start eligible;
-                        // removed voters disappear from both sets in
-                        // the same step. Without preserving surviving
-                        // demotions, a membership change that doesn't
-                        // drop the demoted voter (e.g. adding a new
-                        // voter to grow the cluster) re-promotes the
-                        // demoted voter, opening an
-                        // `unreachable_after`-long window where work
-                        // rendezvouses to the unreachable node.
-                        let new_voters: BTreeSet<NodeId> = m.voter_ids().collect();
-                        let surviving_demotions: BTreeSet<NodeId> = state
-                            .configured_voters
-                            .difference(&state.worker_eligible_voters)
-                            .copied()
-                            .filter(|id| new_voters.contains(id))
-                            .collect();
-                        state.worker_eligible_voters = new_voters
-                            .difference(&surviving_demotions)
-                            .copied()
-                            .collect();
-                        state.configured_voters = new_voters;
-                        self.last_membership = StoredMembership::new(Some(log_id), m);
-                        responses.push(Response::NoOp);
-                    }
-                }
-            }
-        }
-        // Report commit-head advances to the ledger cache first —
-        // synchronous and lossless, unlike the bounded event bus —
-        // so a cache hit racing this apply already compares against
-        // the new watermark.
-        if let Some(manager) = self.ledger_manager.get() {
-            for event in &events {
-                if let NameServiceEvent::LedgerCommitPublished {
-                    ledger_id,
-                    commit_t,
-                    ..
-                } = event
-                {
-                    manager.note_head_advance(ledger_id, *commit_t);
-                }
-            }
-        }
-        // Emit after the state write lock drops so subscribers
-        // can't block apply progress.
-        if let Some(bus) = self.event_bus.as_ref() {
-            for event in events {
-                bus.notify(event);
-            }
-        }
-        if let Some(waiters) = self.waiter_map.as_ref() {
-            for resolution in resolutions {
-                match resolution {
-                    WaiterResolution::Bind {
-                        request_cid,
-                        queue_id,
-                    } => waiters.bind(&request_cid, queue_id),
-                    WaiterResolution::Applied {
-                        queue_id,
-                        commit_id,
-                        commit_t,
-                    } => {
-                        let receipt = self
-                            .staged_receipts
-                            .as_ref()
-                            .and_then(|s| s.take(queue_id))
-                            .unwrap_or(AppliedReceipt::Minimal {
-                                commit_id,
-                                commit_t,
-                            });
-                        waiters.resolve_applied(queue_id, receipt);
-                    }
-                    WaiterResolution::Aborted { queue_id, reason } => {
-                        // The receipt is stale — the queue entry won't
-                        // produce an `ApplyHead` now. Drop it so the
-                        // map doesn't accumulate one entry per poison
-                        // for the lifetime of the process.
-                        if let Some(s) = self.staged_receipts.as_ref() {
-                            s.take(queue_id);
-                        }
-                        waiters.resolve_aborted(queue_id, reason);
-                    }
-                    WaiterResolution::AbortBranch { ref_key, reason } => {
-                        // Admin clear blew away every queue entry on
-                        // the branch — drain their stashed receipts
-                        // too. Same reasoning as the `Aborted` arm:
-                        // no `ApplyHead` will ever consume them.
-                        if let Some(s) = self.staged_receipts.as_ref() {
-                            let _ = s.take_for_ref_key(&ref_key);
-                        }
-                        waiters.abort_all_for_branch(&ref_key, reason);
-                    }
-                }
-            }
-        }
-        if let Some(tx) = self.release_tx.as_ref() {
-            for release in releases {
-                // Receiver dropped → release task already gone
-                // (shutdown window, or the task died). The blob
-                // leaks; log its CID so an operator or offline
-                // sweep can reclaim it.
-                if let Err(tokio::sync::mpsc::error::SendError((ledger_id, cid))) = tx.send(release)
-                {
-                    tracing::warn!(
-                        %ledger_id,
-                        %cid,
-                        "release channel closed; envelope blob will leak"
-                    );
-                }
-            }
-        }
-        Ok(responses)
-    }
-
-    async fn get_snapshot_builder(&mut self) -> Self::SnapshotBuilder {
-        let counter = self.snapshot_counter.fetch_add(1, Ordering::Relaxed) + 1;
-        let state_clone = self.state.read().await.clone();
-        SnapshotBuilder {
-            state: state_clone,
-            last_applied: self.last_applied,
-            last_membership: self.last_membership.clone(),
-            storage: Arc::clone(&self.storage),
-            counter,
-        }
-    }
-
-    async fn begin_receiving_snapshot(
-        &mut self,
-    ) -> Result<Box<Cursor<Vec<u8>>>, StorageError<NodeId>> {
-        Ok(Box::new(Cursor::new(Vec::new())))
-    }
-
-    async fn install_snapshot(
-        &mut self,
-        meta: &SnapshotMeta<NodeId, ClusterNode>,
-        snapshot: Box<Cursor<Vec<u8>>>,
-    ) -> Result<(), StorageError<NodeId>> {
-        let bytes = snapshot.into_inner();
-        let new_state = NameServiceState::from_snapshot(&bytes).map_err(read_state_err)?;
-        // A snapshot install advances heads in bulk without
-        // per-entry applies, so the per-apply watermark report
-        // never fires for them. Capture every restored head here
-        // and report after the state swap, so caches behind the
-        // restored state revalidate.
-        let restored_heads: Vec<(String, i64)> = new_state
-            .refs
-            .iter()
-            .map(|(key, entry)| (key.ledger_id(), entry.t))
-            .collect();
-        let membership_bytes =
-            postcard::to_allocvec(&meta.last_membership).map_err(write_state_err)?;
-
-        self.storage
-            .snapshots()
-            .write(
-                &OurSnapshotMeta {
-                    id: OurSnapshotId::new(&meta.snapshot_id),
-                    last_applied: meta.last_log_id.map(from_openraft_log_id),
-                    membership: membership_bytes,
-                },
-                bytes,
-            )
-            .await
-            .map_err(|e| snapshot_err(ErrorVerb::Write, e))?;
-
-        *self.state.write().await = new_state;
-        self.last_applied = meta.last_log_id;
-        self.last_membership = meta.last_membership.clone();
-
-        if let Some(manager) = self.ledger_manager.get() {
-            for (ledger_id, commit_t) in &restored_heads {
-                manager.note_head_advance(ledger_id, *commit_t);
-            }
-        }
-
-        // The snapshot replaces the state machine wholesale. Any
-        // queue_id tracked by this node's waiter map or staged-
-        // receipt map belonged to the prior-leader state, and the
-        // new state may or may not contain it — neither parked
-        // proposers nor buffered receipts can be trusted across the
-        // install. Resolve waiters as aborted so callers stop
-        // waiting; drop the stash so memory doesn't leak.
-        if let Some(waiters) = self.waiter_map.as_ref() {
-            waiters.drain_all_with(AbortReason::SnapshotInstalled);
-        }
-        if let Some(stash) = self.staged_receipts.as_ref() {
-            stash.clear_all();
-        }
-        Ok(())
-    }
-
-    async fn get_current_snapshot(
-        &mut self,
-    ) -> Result<Option<Snapshot<TypeConfig>>, StorageError<NodeId>> {
-        let current = self
-            .storage
-            .snapshots()
-            .current()
-            .await
-            .map_err(|e| snapshot_err(ErrorVerb::Read, e))?;
-        let Some((our_meta, data)) = current else {
-            return Ok(None);
-        };
-        let last_membership: StoredMembership<NodeId, ClusterNode> =
-            postcard::from_bytes(&our_meta.membership).map_err(read_state_err)?;
-        let meta = SnapshotMeta {
-            last_log_id: our_meta.last_applied.map(to_openraft_log_id),
-            last_membership,
-            snapshot_id: our_meta.id.as_str().to_string(),
-        };
-        Ok(Some(Snapshot {
-            meta,
-            snapshot: Box::new(Cursor::new(data)),
-        }))
-    }
-}
-
-/// Snapshot builder produced by
-/// [`StateMachineAdapter::get_snapshot_builder`]. Carries a cloned
-/// view of state-machine state so the snapshot reflects the moment
-/// the builder was obtained, not "now."
-pub struct SnapshotBuilder<S>
-where
-    S: RaftStorage,
-{
-    state: NameServiceState,
-    last_applied: Option<LogId<NodeId>>,
-    last_membership: StoredMembership<NodeId, ClusterNode>,
-    storage: Arc<S>,
-    counter: u64,
-}
-
-impl<S> RaftSnapshotBuilder<TypeConfig> for SnapshotBuilder<S>
-where
-    S: RaftStorage,
-{
-    async fn build_snapshot(&mut self) -> Result<Snapshot<TypeConfig>, StorageError<NodeId>> {
-        let bytes = self.state.to_snapshot().map_err(write_state_err)?;
-        let last_index = self.last_applied.map(|id| id.index).unwrap_or(0);
-        let snapshot_id = format!("snap-{}-{}", last_index, self.counter);
-
-        let membership_bytes =
-            postcard::to_allocvec(&self.last_membership).map_err(write_state_err)?;
-        self.storage
-            .snapshots()
-            .write(
-                &OurSnapshotMeta {
-                    id: OurSnapshotId::new(&snapshot_id),
-                    last_applied: self.last_applied.map(from_openraft_log_id),
-                    membership: membership_bytes,
-                },
-                bytes.clone(),
-            )
-            .await
-            .map_err(|e| snapshot_err(ErrorVerb::Write, e))?;
-
-        let meta = SnapshotMeta {
-            last_log_id: self.last_applied,
-            last_membership: self.last_membership.clone(),
-            snapshot_id,
-        };
-        Ok(Snapshot {
-            meta,
-            snapshot: Box::new(Cursor::new(bytes)),
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Outcomes are delivered synchronously by `apply`; this only
-    /// bounds a genuine failure.
-    const WAIT: std::time::Duration = std::time::Duration::from_secs(5);
-    use crate::raft::state_machine::NewLedger;
-    use crate::raft::storage::memory::MemoryRaftStorage;
-    use crate::raft::Command as RaftCommand;
-    use fluree_db_api::{ContentId, ContentKind};
-    use openraft::{CommittedLeaderId, LogId};
-
-    fn cid(seed: u8) -> ContentId {
-        ContentId::new(ContentKind::Commit, &[seed])
-    }
-
-    /// `drain_releases` must forward a keyless entry's envelope
-    /// (paired with the response's ledger_id) so the per-node
-    /// release task frees it, and forward nothing when the entry
-    /// kept its envelope for later idempotency eviction.
-    #[test]
-    fn drain_releases_forwards_keyless_apply_envelopes() {
-        let mut applied = Response::HeadApplied {
-            ledger_id: "test/db:main".into(),
-            commit_id: cid(1),
-            commit_t: 1,
-            released_envelope: Some(cid(9)),
-        };
-        assert_eq!(
-            drain_releases(&mut applied),
-            vec![("test/db:main".to_string(), cid(9))]
-        );
-
-        let mut keyed = Response::HeadApplied {
-            ledger_id: "test/db:main".into(),
-            commit_id: cid(1),
-            commit_t: 1,
-            released_envelope: None,
-        };
-        assert!(drain_releases(&mut keyed).is_empty());
-
-        let mut poisoned = Response::Poisoned {
-            ledger_id: "test/db:main".into(),
-            queue_id: 7,
-            reason: crate::raft::state_machine::PoisonReason::BodyMalformed { error: "x".into() },
-            released_envelope: Some(cid(3)),
-        };
-        assert_eq!(
-            drain_releases(&mut poisoned),
-            vec![("test/db:main".to_string(), cid(3))]
-        );
-    }
-
-    fn log_id(term: u64, index: u64) -> LogId<NodeId> {
-        LogId {
-            leader_id: CommittedLeaderId::new(term, 0),
-            index,
-        }
-    }
-
-    fn create_ledger_entry(index: u64, ledger_id: &str) -> Entry<TypeConfig> {
-        Entry {
-            log_id: log_id(1, index),
-            payload: EntryPayload::Normal(RaftCommand::CreateLedger(NewLedger {
-                ledger_id: ledger_id.into(),
-                branch: "main".into(),
-                created_at_millis: 1_000,
-            })),
-        }
-    }
-
-    #[tokio::test]
-    async fn apply_routes_create_ledger_to_state_machine() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-        let responses = sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        assert_eq!(responses.len(), 1);
-        assert!(matches!(responses[0], Response::Created { .. }));
-        assert!(sm
-            .shared_state()
-            .read()
-            .await
-            .ledgers
-            .contains_key("test/db"));
-        let (applied, _) = sm.applied_state().await.unwrap();
-        assert_eq!(applied, Some(log_id(1, 1)));
-    }
-
-    #[tokio::test]
-    async fn blank_entry_is_noop_but_advances_last_applied() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-        let blank = Entry {
-            log_id: log_id(1, 5),
-            payload: EntryPayload::Blank,
-        };
-        let responses = sm.apply([blank]).await.unwrap();
-        assert_eq!(responses, vec![Response::NoOp]);
-        let (applied, _) = sm.applied_state().await.unwrap();
-        assert_eq!(applied, Some(log_id(1, 5)));
-    }
-
-    fn cluster_node(id: NodeId) -> ClusterNode {
-        ClusterNode::new(
-            format!("http://node-{id}:9090/raft"),
-            format!("http://node-{id}:8080"),
-        )
-    }
-
-    fn membership_entry(index: u64, voters: &[NodeId]) -> Entry<TypeConfig> {
-        let voter_set: BTreeSet<NodeId> = voters.iter().copied().collect();
-        let nodes: std::collections::BTreeMap<NodeId, ClusterNode> =
-            voters.iter().map(|&id| (id, cluster_node(id))).collect();
-        Entry {
-            log_id: log_id(1, index),
-            payload: EntryPayload::Membership(openraft::Membership::new(vec![voter_set], nodes)),
-        }
-    }
-
-    #[tokio::test]
-    async fn membership_apply_mirrors_voter_set_into_state() {
-        // First membership-apply sets `configured_voters` from
-        // empty and seeds every voter as eligible — the shape the
-        // worker supervisor's rendezvous expects at cluster boot.
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-
-        sm.apply([membership_entry(1, &[1, 2, 3])]).await.unwrap();
-
-        let shared = sm.shared_state();
-        let state = shared.read().await;
-        let expected: BTreeSet<NodeId> = [1, 2, 3].into_iter().collect();
-        assert_eq!(state.configured_voters, expected);
-        assert_eq!(state.worker_eligible_voters, expected);
-    }
-
-    #[tokio::test]
-    async fn membership_change_removing_demoted_voter_drops_it_from_both_sets() {
-        // A prior membership-apply seeded {1,2,3}; the leader's
-        // monitor then demoted 2. The next membership-apply that
-        // adds 4 and removes 2 should snap both sets to {1,3,4}
-        // — newly-configured voters start eligible, demoted-then-
-        // removed voters disappear from both sets.
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-
-        sm.apply([membership_entry(1, &[1, 2, 3])]).await.unwrap();
-        let shared = sm.shared_state();
-        // Mid-life demotion of 2.
-        {
-            let mut state = shared.write().await;
-            state.worker_eligible_voters.remove(&2);
-        }
-
-        // Membership change: drop 2, add 4.
-        sm.apply([membership_entry(2, &[1, 3, 4])]).await.unwrap();
-
-        let state = shared.read().await;
-        let expected: BTreeSet<NodeId> = [1, 3, 4].into_iter().collect();
-        assert_eq!(state.configured_voters, expected);
-        // 4 starts eligible alongside the survivors, 2 is gone.
-        assert_eq!(state.worker_eligible_voters, expected);
-    }
-
-    #[tokio::test]
-    async fn membership_change_preserves_demotion_for_voter_that_survives() {
-        // A prior membership-apply seeded {1,2,3}; the leader's
-        // monitor demoted 2 (still unreachable). The next
-        // membership-apply *adds* 4 without dropping 2 — common
-        // case when growing the cluster while a node is down.
-        // 2's demotion must survive the membership change:
-        // worker_eligible_voters should be {1,3,4}, not {1,2,3,4}.
-        // Without this guard, work would rendezvous to the
-        // unreachable 2 until the monitor re-demoted it
-        // ~`unreachable_after` later.
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-
-        sm.apply([membership_entry(1, &[1, 2, 3])]).await.unwrap();
-        let shared = sm.shared_state();
-        // Mid-life demotion of 2.
-        {
-            let mut state = shared.write().await;
-            state.worker_eligible_voters.remove(&2);
-        }
-
-        // Membership change: add 4, keep 2 in the voter set.
-        sm.apply([membership_entry(2, &[1, 2, 3, 4])])
-            .await
-            .unwrap();
-
-        let state = shared.read().await;
-        let expected_configured: BTreeSet<NodeId> = [1, 2, 3, 4].into_iter().collect();
-        let expected_eligible: BTreeSet<NodeId> = [1, 3, 4].into_iter().collect();
-        assert_eq!(state.configured_voters, expected_configured);
-        assert_eq!(
-            state.worker_eligible_voters, expected_eligible,
-            "demotion of 2 must survive the membership change; 4 starts eligible"
-        );
-    }
-
-    #[tokio::test]
-    async fn membership_change_starts_newly_added_voters_eligible() {
-        // No prior demotions, just a cluster-growth scenario:
-        // {1,2,3} → {1,2,3,4,5}. Both new voters start eligible.
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-
-        sm.apply([membership_entry(1, &[1, 2, 3])]).await.unwrap();
-        sm.apply([membership_entry(2, &[1, 2, 3, 4, 5])])
-            .await
-            .unwrap();
-
-        let shared = sm.shared_state();
-        let state = shared.read().await;
-        let expected: BTreeSet<NodeId> = [1, 2, 3, 4, 5].into_iter().collect();
-        assert_eq!(state.configured_voters, expected);
-        assert_eq!(state.worker_eligible_voters, expected);
-    }
-
-    /// Direct head seed used by event-bus tests that need a populated
-    /// branch head as setup, but don't care which mechanism set it
-    /// (and don't want a queue-path `LedgerCommitPublished` event to
-    /// leak into the test's bus draining).
-    async fn seed_branch_head<S: super::RaftStorage>(
-        sm: &StateMachineAdapter<S>,
-        ledger_id: &str,
-        branch: &str,
-        head: ContentId,
-        t: i64,
-    ) {
-        let mut guard = sm.shared_state().write_owned().await;
-        if let Some(ledger) = guard.ledgers.get_mut(ledger_id) {
-            if !ledger.branches.iter().any(|b| b == branch) {
-                ledger.branches.push(branch.to_string());
-            }
-        }
-        let ref_key = crate::raft::state_machine::RefKey::new(ledger_id, branch);
-        let (prior_index, prior_source, prior_branches) = guard
-            .refs
-            .get(&ref_key)
-            .map(|r| (r.index.clone(), r.source_branch.clone(), r.branches))
-            .unwrap_or_default();
-        guard.refs.insert(
-            ref_key,
-            crate::raft::state_machine::RefEntry {
-                head,
-                t,
-                last_advanced_at_millis: 2_000,
-                last_advanced_index: 1,
-                index: prior_index,
-                source_branch: prior_source,
-                branches: prior_branches,
-            },
-        );
-    }
-
-    #[tokio::test]
-    async fn snapshot_build_persists_and_get_current_round_trips() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(Arc::clone(&storage));
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-
-        let mut builder = sm.get_snapshot_builder().await;
-        let snap = builder.build_snapshot().await.unwrap();
-        assert_eq!(snap.meta.last_log_id, Some(log_id(1, 1)));
-
-        let current = sm.get_current_snapshot().await.unwrap().unwrap();
-        assert_eq!(current.meta.snapshot_id, snap.meta.snapshot_id);
-    }
-
-    #[tokio::test]
-    async fn apply_emits_retracted_event_on_fresh_retract() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let bus = Arc::new(LedgerEventBus::new(16));
-        let mut sm = StateMachineAdapter::new(storage).with_event_bus(Arc::clone(&bus));
-        let mut sub = bus.subscribe(fluree_db_nameservice::SubscriptionScope::All);
-
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        sm.apply([Entry {
-            log_id: log_id(1, 2),
-            payload: EntryPayload::Normal(RaftCommand::RetractLedger {
-                ledger_id: "test/db".into(),
-                branch: "main".into(),
-                applied_at_millis: 0,
-            }),
-        }])
-        .await
-        .unwrap();
-
-        match sub.receiver.try_recv().expect("retracted event") {
-            NameServiceEvent::LedgerRetracted { ledger_id } => {
-                assert_eq!(ledger_id, "test/db:main");
-            }
-            other => panic!("expected LedgerRetracted, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn apply_emits_nothing_on_already_retracted() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let bus = Arc::new(LedgerEventBus::new(16));
-        let mut sm = StateMachineAdapter::new(storage).with_event_bus(Arc::clone(&bus));
-        let mut sub = bus.subscribe(fluree_db_nameservice::SubscriptionScope::All);
-
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        sm.apply([Entry {
-            log_id: log_id(1, 2),
-            payload: EntryPayload::Normal(RaftCommand::RetractLedger {
-                ledger_id: "test/db".into(),
-                branch: "main".into(),
-                applied_at_millis: 0,
-            }),
-        }])
-        .await
-        .unwrap();
-        let _ = sub.receiver.try_recv().expect("first retract emits");
-
-        sm.apply([Entry {
-            log_id: log_id(1, 3),
-            payload: EntryPayload::Normal(RaftCommand::RetractLedger {
-                ledger_id: "test/db".into(),
-                branch: "main".into(),
-                applied_at_millis: 0,
-            }),
-        }])
-        .await
-        .unwrap();
-        assert!(
-            sub.receiver.try_recv().is_err(),
-            "idempotent retract should not emit"
-        );
-    }
-
-    #[tokio::test]
-    async fn apply_emits_retracted_event_on_purge_of_known_branch() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let bus = Arc::new(LedgerEventBus::new(16));
-        let mut sm = StateMachineAdapter::new(storage).with_event_bus(Arc::clone(&bus));
-        let mut sub = bus.subscribe(fluree_db_nameservice::SubscriptionScope::All);
-
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        sm.apply([Entry {
-            log_id: log_id(1, 2),
-            payload: EntryPayload::Normal(RaftCommand::PurgeBranch {
-                ledger_id: "test/db".into(),
-                branch: "main".into(),
-                applied_at_millis: 0,
-            }),
-        }])
-        .await
-        .unwrap();
-
-        match sub.receiver.try_recv().expect("purge event") {
-            NameServiceEvent::LedgerRetracted { ledger_id } => {
-                assert_eq!(ledger_id, "test/db:main");
-            }
-            other => panic!("expected LedgerRetracted, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn apply_emits_nothing_on_purge_of_missing_branch() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let bus = Arc::new(LedgerEventBus::new(16));
-        let mut sm = StateMachineAdapter::new(storage).with_event_bus(Arc::clone(&bus));
-        let mut sub = bus.subscribe(fluree_db_nameservice::SubscriptionScope::All);
-
-        sm.apply([Entry {
-            log_id: log_id(1, 1),
-            payload: EntryPayload::Normal(RaftCommand::PurgeBranch {
-                ledger_id: "ghost".into(),
-                branch: "main".into(),
-                applied_at_millis: 0,
-            }),
-        }])
-        .await
-        .unwrap();
-        assert!(
-            sub.receiver.try_recv().is_err(),
-            "purge of unknown branch should not emit"
-        );
-    }
-
-    #[tokio::test]
-    async fn apply_emits_commit_event_on_create_branch() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let bus = Arc::new(LedgerEventBus::new(16));
-        let mut sm = StateMachineAdapter::new(storage).with_event_bus(Arc::clone(&bus));
-        let mut sub = bus.subscribe(fluree_db_nameservice::SubscriptionScope::All);
-
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        seed_branch_head(&sm, "test/db", "main", cid(7), 10).await;
-
-        sm.apply([Entry {
-            log_id: log_id(1, 3),
-            payload: EntryPayload::Normal(RaftCommand::CreateBranch(
-                crate::raft::state_machine::NewBranch {
-                    ledger_id: "test/db".into(),
-                    branch: "feature".into(),
-                    source_branch: "main".into(),
-                    at_commit: None,
-                    applied_at_millis: 3_000,
-                },
-            )),
-        }])
-        .await
-        .unwrap();
-
-        match sub.receiver.try_recv().expect("create-branch event") {
-            NameServiceEvent::LedgerCommitPublished {
-                ledger_id,
-                commit_id,
-                commit_t,
-            } => {
-                assert_eq!(ledger_id, "test/db:feature");
-                assert_eq!(commit_id, cid(7));
-                assert_eq!(commit_t, 10);
-            }
-            other => panic!("expected LedgerCommitPublished, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn apply_emits_retracted_event_on_drop_branch() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let bus = Arc::new(LedgerEventBus::new(16));
-        let mut sm = StateMachineAdapter::new(storage).with_event_bus(Arc::clone(&bus));
-        let mut sub = bus.subscribe(fluree_db_nameservice::SubscriptionScope::All);
-
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        seed_branch_head(&sm, "test/db", "main", cid(7), 10).await;
-        sm.apply([Entry {
-            log_id: log_id(1, 3),
-            payload: EntryPayload::Normal(RaftCommand::CreateBranch(
-                crate::raft::state_machine::NewBranch {
-                    ledger_id: "test/db".into(),
-                    branch: "feature".into(),
-                    source_branch: "main".into(),
-                    at_commit: None,
-                    applied_at_millis: 3_000,
-                },
-            )),
-        }])
-        .await
-        .unwrap();
-        // Drain the create-branch event.
-        let _ = sub.receiver.try_recv().expect("create-branch event");
-
-        sm.apply([Entry {
-            log_id: log_id(1, 4),
-            payload: EntryPayload::Normal(RaftCommand::DropBranch {
-                ledger_id: "test/db".into(),
-                branch: "feature".into(),
-                applied_at_millis: 0,
-            }),
-        }])
-        .await
-        .unwrap();
-
-        match sub.receiver.try_recv().expect("drop-branch event") {
-            NameServiceEvent::LedgerRetracted { ledger_id } => {
-                assert_eq!(ledger_id, "test/db:feature");
-            }
-            other => panic!("expected LedgerRetracted, got {other:?}"),
-        }
-    }
-
-    /// A real `LedgerManager` over memory backends, for asserting
-    /// watermark reports through [`LedgerManager::head_watermark`].
-    fn memory_ledger_manager() -> Arc<LedgerManager> {
-        use fluree_db_api::{LedgerManagerConfig, NameServiceMode};
-        use fluree_db_core::{MemoryStorage, StorageBackend};
-        use fluree_db_nameservice::memory::MemoryNameService;
-
-        Arc::new(LedgerManager::new(
-            StorageBackend::Managed(Arc::new(MemoryStorage::new())),
-            NameServiceMode::ReadWrite(Arc::new(MemoryNameService::new())),
-            LedgerManagerConfig::default(),
-        ))
-    }
-
-    #[tokio::test]
-    async fn apply_reports_commit_head_advances_to_ledger_manager() {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut sm = StateMachineAdapter::new(storage);
-        let manager = memory_ledger_manager();
-        sm.ledger_manager_cell()
-            .set(Arc::clone(&manager))
-            .ok()
-            .expect("cell empty");
-
-        sm.apply([create_ledger_entry(1, "test/db")]).await.unwrap();
-        seed_branch_head(&sm, "test/db", "main", cid(7), 10).await;
-        // CreateBranch emits `LedgerCommitPublished` for the new
-        // branch at the source head — the cache's watermark must
-        // reflect the same advance the event carries.
-        sm.apply([Entry {
-            log_id: log_id(1, 3),
-            payload: EntryPayload::Normal(RaftCommand::CreateBranch(
-                crate::raft::state_machine::NewBranch {
-                    ledger_id: "test/db".into(),
-                    branch: "feature".into(),
-                    source_branch: "main".into(),
-                    at_commit: None,
-                    applied_at_millis: 3_000,
-                },
-            )),
-        }])
-        .await
-        .unwrap();
-
-        assert_eq!(manager.head_watermark("test/db:feature"), Some(10));
-    }
-
-    #[tokio::test]
-    async fn install_snapshot_reports_restored_heads_to_ledger_manager() {
-        let source_storage = Arc::new(MemoryRaftStorage::new());
-        let mut source = StateMachineAdapter::new(Arc::clone(&source_storage));
-        source
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        seed_branch_head(&source, "test/db", "main", cid(7), 42).await;
-        let mut builder = source.get_snapshot_builder().await;
-        let snap = builder.build_snapshot().await.unwrap();
-
-        // A snapshot install moves heads in bulk with no per-entry
-        // applies; the sweep must report every restored head or
-        // caches keyed on the old state never revalidate.
-        let target_storage = Arc::new(MemoryRaftStorage::new());
-        let mut target = StateMachineAdapter::new(target_storage);
-        let manager = memory_ledger_manager();
-        target
-            .ledger_manager_cell()
-            .set(Arc::clone(&manager))
-            .ok()
-            .expect("cell empty");
-
-        target
-            .install_snapshot(&snap.meta, snap.snapshot)
-            .await
-            .unwrap();
-
-        assert_eq!(manager.head_watermark("test/db:main"), Some(42));
-    }
-
-    #[tokio::test]
-    async fn install_snapshot_replaces_state_and_persists() {
-        let source_storage = Arc::new(MemoryRaftStorage::new());
-        let mut source = StateMachineAdapter::new(Arc::clone(&source_storage));
-        source
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        let mut builder = source.get_snapshot_builder().await;
-        let snap = builder.build_snapshot().await.unwrap();
-
-        let target_storage = Arc::new(MemoryRaftStorage::new());
-        let mut target = StateMachineAdapter::new(Arc::clone(&target_storage));
-        target
-            .install_snapshot(&snap.meta, snap.snapshot)
-            .await
-            .unwrap();
-
-        assert!(target
-            .shared_state()
-            .read()
-            .await
-            .ledgers
-            .contains_key("test/db"));
-        let (applied, _) = target.applied_state().await.unwrap();
-        assert_eq!(applied, Some(log_id(1, 1)));
-    }
-
-    #[tokio::test]
-    async fn install_snapshot_drains_waiters_and_stash() {
-        use crate::raft::staged_receipt::{StagedReceiptMap, TransactApplied};
-
-        // Source adapter builds a snapshot from a fresh state.
-        let source_storage = Arc::new(MemoryRaftStorage::new());
-        let mut source = StateMachineAdapter::new(Arc::clone(&source_storage));
-        source
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        let mut builder = source.get_snapshot_builder().await;
-        let snap = builder.build_snapshot().await.unwrap();
-
-        // Target adapter has both maps wired and prior-leader work
-        // tracked in them: a parked waiter, a buffered Applied, and
-        // a stashed receipt — all keyed on queue_ids the snapshot
-        // does not (and cannot) recognize.
-        let target_storage = Arc::new(MemoryRaftStorage::new());
-        let waiter_map = Arc::new(WaiterMap::new());
-        let staged = Arc::new(StagedReceiptMap::new());
-        let mut target = StateMachineAdapter::new(target_storage)
-            .with_waiter_map(Arc::clone(&waiter_map))
-            .with_staged_receipts(Arc::clone(&staged));
-
-        let mut bound = waiter_map.arm_bound(cid(210), RefKey::new("test/db", "main"), 100);
-        // A submission that was proposed but whose enqueue has not
-        // applied here yet — still an interest, no queue_id.
-        let mut unbound = waiter_map.arm(cid(211), RefKey::new("test/db", "main"));
-        // A terminal apply for an id no local proposer is waiting on —
-        // every such apply on a follower. It must leave nothing behind.
-        waiter_map.resolve_applied(
-            101,
-            AppliedReceipt::Minimal {
-                commit_id: cid(7),
-                commit_t: 5,
-            },
-        );
-        staged.stash(
-            100,
-            RefKey::new("test/db", "main"),
-            AppliedReceipt::Transact(TransactApplied {
-                commit_id: cid(42),
-                commit_t: 10,
-                flake_count: 0,
-                tally: None,
-            }),
-        );
-        assert_eq!(
-            waiter_map.len(),
-            2,
-            "one bound waiter and one armed interest; the unmatched resolve tracks nothing",
-        );
-        assert_eq!(staged.len(), 1);
-
-        target
-            .install_snapshot(&snap.meta, snap.snapshot)
-            .await
-            .unwrap();
-
-        // Both the bound waiter and the armed interest are abandoned:
-        // the snapshot replaced the state wholesale, so neither the
-        // queue entry nor the pending enqueue can be trusted to exist.
-        assert!(matches!(
-            bound.wait(WAIT).await.unwrap(),
-            WaiterOutcome::Aborted(AbortReason::SnapshotInstalled)
-        ));
-        assert!(matches!(
-            unbound.wait(WAIT).await.unwrap(),
-            WaiterOutcome::Aborted(AbortReason::SnapshotInstalled)
-        ));
-        assert!(waiter_map.is_empty());
-        // Stashed receipts cleared.
-        assert!(staged.is_empty());
-    }
-
-    // ====================================================================
-    // Waiter-map resolution
-    // ====================================================================
-
-    use crate::raft::state_machine::{
-        BodyKind, EntryPoisoning, PoisonReason, QueueSubmission, ResetHeadSnapshot, StagedHead,
-    };
-    use crate::raft::waiter::{AbortReason, WaiterMap, WaiterOutcome};
-
-    fn enqueue_entry(index: u64, ledger_id: &str, branch: &str) -> Entry<TypeConfig> {
-        Entry {
-            log_id: log_id(1, index),
-            payload: EntryPayload::Normal(RaftCommand::EnqueueCommand(QueueSubmission {
-                ledger_id: ledger_id.into(),
-                branch: branch.into(),
-                idempotency: None,
-                request_cid: cid(0),
-                body_cid: cid(0),
-                body_kind: BodyKind::JsonLdInsert,
-                applied_at_millis: 1_500,
-            })),
-        }
-    }
-
-    fn apply_head_entry(
-        index: u64,
-        ledger_id: &str,
-        branch: &str,
-        queue_id: u64,
-        commit: ContentId,
-        commit_t: i64,
-    ) -> Entry<TypeConfig> {
-        Entry {
-            log_id: log_id(1, index),
-            payload: EntryPayload::Normal(RaftCommand::ApplyHead(StagedHead {
-                ledger_id: ledger_id.into(),
-                branch: branch.into(),
-                queue_id,
-                commit_id: commit,
-                commit_t,
-                applied_at_millis: 2_000,
-                tally: None,
-                flake_count: 0,
-            })),
-        }
-    }
-
-    fn poison_entry(
-        index: u64,
-        ledger_id: &str,
-        branch: &str,
-        queue_id: u64,
-        reason: PoisonReason,
-    ) -> Entry<TypeConfig> {
-        Entry {
-            log_id: log_id(1, index),
-            payload: EntryPayload::Normal(RaftCommand::PoisonQueueEntry(EntryPoisoning {
-                ledger_id: ledger_id.into(),
-                branch: branch.into(),
-                queue_id,
-                reason,
-                applied_at_millis: 2_000,
-            })),
-        }
-    }
-
-    fn drop_branch_entry(index: u64, ledger_id: &str, branch: &str) -> Entry<TypeConfig> {
-        Entry {
-            log_id: log_id(1, index),
-            payload: EntryPayload::Normal(RaftCommand::DropBranch {
-                ledger_id: ledger_id.into(),
-                branch: branch.into(),
-                applied_at_millis: 0,
-            }),
-        }
-    }
-
-    async fn adapter_with_waiters() -> (StateMachineAdapter<MemoryRaftStorage>, Arc<WaiterMap>) {
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let waiter_map = Arc::new(WaiterMap::new());
-        let adapter = StateMachineAdapter::new(storage).with_waiter_map(Arc::clone(&waiter_map));
-        (adapter, waiter_map)
-    }
-
-    #[tokio::test]
-    async fn apply_head_resolves_waiter_with_applied_outcome() {
-        let (mut adapter, waiters) = adapter_with_waiters().await;
-        adapter
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        adapter
-            .apply([enqueue_entry(2, "test/db", "main")])
-            .await
-            .unwrap();
-
-        let mut rx = waiters.arm_bound(cid(20), RefKey::new("test/db", "main"), 0);
-        adapter
-            .apply([apply_head_entry(3, "test/db", "main", 0, cid(42), 10)])
-            .await
-            .unwrap();
-
-        // No StagedReceiptMap is configured on the adapter, so the
-        // resolution falls back to Minimal — confirming the absent-
-        // entry path delivers commit_id / commit_t without panicking.
-        match rx.wait(WAIT).await.expect("receive") {
-            WaiterOutcome::Applied(AppliedReceipt::Minimal {
-                commit_id,
-                commit_t,
-            }) => {
-                assert_eq!(commit_id, cid(42));
-                assert_eq!(commit_t, 10);
-            }
-            other => panic!("expected Applied(Minimal), got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn apply_head_reads_stashed_receipt_when_present() {
-        use crate::raft::staged_receipt::{StagedReceiptMap, TransactApplied};
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let waiter_map = Arc::new(WaiterMap::new());
-        let staged = Arc::new(StagedReceiptMap::new());
-        let mut adapter = StateMachineAdapter::new(storage)
-            .with_waiter_map(Arc::clone(&waiter_map))
-            .with_staged_receipts(Arc::clone(&staged));
-        adapter
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        adapter
-            .apply([enqueue_entry(2, "test/db", "main")])
-            .await
-            .unwrap();
-
-        let mut rx = waiter_map.arm_bound(cid(20), RefKey::new("test/db", "main"), 0);
-        staged.stash(
-            0,
-            RefKey::new("test/db", "main"),
-            AppliedReceipt::Transact(TransactApplied {
-                commit_id: cid(42),
-                commit_t: 10,
-                flake_count: 0,
-                tally: None,
-            }),
-        );
-        adapter
-            .apply([apply_head_entry(3, "test/db", "main", 0, cid(42), 10)])
-            .await
-            .unwrap();
-
-        match rx.wait(WAIT).await.expect("receive") {
-            WaiterOutcome::Applied(AppliedReceipt::Transact(r)) => {
-                assert_eq!(r.commit_id, cid(42));
-                assert_eq!(r.commit_t, 10);
-            }
-            other => panic!("expected Applied(Transact), got {other:?}"),
-        }
-        assert_eq!(staged.len(), 0, "adapter must take from the map");
-    }
-
-    #[tokio::test]
-    async fn poison_resolves_waiter_with_aborted_poisoned() {
-        let (mut adapter, waiters) = adapter_with_waiters().await;
-        adapter
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        adapter
-            .apply([enqueue_entry(2, "test/db", "main")])
-            .await
-            .unwrap();
-
-        let mut rx = waiters.arm_bound(cid(20), RefKey::new("test/db", "main"), 0);
-        adapter
-            .apply([poison_entry(
-                3,
-                "test/db",
-                "main",
-                0,
-                PoisonReason::BodyMalformed {
-                    error: "bad turtle".into(),
-                },
-            )])
-            .await
-            .unwrap();
-
-        match rx.wait(WAIT).await.expect("receive") {
-            WaiterOutcome::Aborted(AbortReason::Poisoned(PoisonReason::BodyMalformed {
-                error,
-            })) => assert_eq!(error, "bad turtle"),
-            other => panic!("expected Poisoned, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn drop_branch_resolves_every_pending_waiter_on_that_branch() {
-        let (mut adapter, waiters) = adapter_with_waiters().await;
-        adapter
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        adapter
-            .apply([enqueue_entry(2, "test/db", "main")])
-            .await
-            .unwrap();
-        adapter
-            .apply([enqueue_entry(3, "test/db", "main")])
-            .await
-            .unwrap();
-
-        let mut rx_a = waiters.arm_bound(cid(20), RefKey::new("test/db", "main"), 0);
-        let mut rx_b = waiters.arm_bound(cid(21), RefKey::new("test/db", "main"), 1);
-
-        adapter
-            .apply([drop_branch_entry(4, "test/db", "main")])
-            .await
-            .unwrap();
-
-        assert!(matches!(
-            rx_a.wait(WAIT).await.unwrap(),
-            WaiterOutcome::Aborted(AbortReason::BranchDropped)
-        ));
-        assert!(matches!(
-            rx_b.wait(WAIT).await.unwrap(),
-            WaiterOutcome::Aborted(AbortReason::BranchDropped)
-        ));
-    }
-
-    #[tokio::test]
-    async fn reset_head_resolves_waiter_with_branch_head_reset() {
-        let (mut adapter, waiters) = adapter_with_waiters().await;
-        adapter
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        adapter
-            .apply([enqueue_entry(2, "test/db", "main")])
-            .await
-            .unwrap();
-
-        let mut rx = waiters.arm_bound(cid(20), RefKey::new("test/db", "main"), 0);
-        adapter
-            .apply([Entry {
-                log_id: log_id(1, 3),
-                payload: EntryPayload::Normal(RaftCommand::ResetHead {
-                    ledger_id: "test/db".into(),
-                    branch: "main".into(),
-                    snapshot: ResetHeadSnapshot {
-                        commit_head_id: None,
-                        commit_t: 0,
-                        index_head_id: None,
-                        index_t: 0,
-                    },
-                    applied_at_millis: 0,
-                }),
-            }])
-            .await
-            .unwrap();
-
-        assert!(matches!(
-            rx.wait(WAIT).await.unwrap(),
-            WaiterOutcome::Aborted(AbortReason::BranchHeadReset)
-        ));
-    }
-
-    #[tokio::test]
-    async fn apply_without_waiter_map_is_silent() {
-        // No waiter_map configured — the adapter should still apply
-        // and respond normally without trying to resolve anything.
-        let storage = Arc::new(MemoryRaftStorage::new());
-        let mut adapter = StateMachineAdapter::new(storage);
-        adapter
-            .apply([create_ledger_entry(1, "test/db")])
-            .await
-            .unwrap();
-        // No assertions beyond "didn't panic" — the absence of a
-        // waiter handle should be benign.
-    }
-}
+//! Build one with [`NameServiceObserver`] carrying whichever effect
+//! sinks this node has:
+//!
+//! ```ignore
+//! let observer = NameServiceObserver::new()
+//!     .with_event_bus(bus)
+//!     .with_waiter_map(waiters);
+//! let sm = StateMachineAdapter::open(storage, observer).await?;
+//! ```
+
+pub use super::app::{Effect, NameServiceApp, NameServiceObserver, SharedState};
+
+/// The nameservice's openraft state machine over an
+/// `Arc<S: RaftStorage>`.
+pub type StateMachineAdapter<S> =
+    fluree_raft_core::state_machine::StateMachineAdapter<NameServiceApp, NameServiceObserver, S>;
